requirements.txt
torch
torchvision
PyYAML
Pillow
numpy